refactor: move EditExerciseScreen to new module
