+++ ui/screens/__init__.py
@@ -1,18 +1,16 @@
 """UI screen modules for WorkoutApp."""
 
 from .exercise_library import ExerciseLibraryScreen
-
-__all__ = ["ExerciseLibraryScreen"]
 from .metric_input_screen import MetricInputScreen
-
-__all__ = ["MetricInputScreen"]
-
-"""Screen module exports."""
-
 from .workout_active_screen import WorkoutActiveScreen
-
-__all__ = ["WorkoutActiveScreen"]
-
 from .preset_detail_screen import PresetDetailScreen
+from .edit_exercise_screen import EditExerciseScreen
 
+__all__ = [
+    "ExerciseLibraryScreen",
+    "MetricInputScreen",
+    "WorkoutActiveScreen",
+    "PresetDetailScreen",
+    "EditExerciseScreen",
+]
